requirements.txt
torch>=1.9.0
torchvision>=0.10.0
flask>=2.0.0
pillow>=8.0.0
werkzeug>=2.0.0
timm>=0.6.0
numpy>=1.21.0